refactor: LottoGenerator model로 이동
