+++ src/main/java/model/LottoGenerator.java
@@ -1,4 +1,4 @@
-package generator;
+package model;
 
 import model.Lotto;
