+++ src/main/java/model/Lottos.java
@@ -1,7 +1,5 @@
 package model;
 
-import generator.LottoGenerator;
-
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
